Refactor and cleanup - add restHelper initial implementation
