+++ pkg/teamcity/rest_helper.go
@@ -1,0 +1,40 @@
+package teamcity
+
+import (
+	"fmt"
+	"io/ioutil"
+	"net/http"
+
+	"github.com/dghubble/sling"
+)
+
+type restHelper struct {
+	httpClient *http.Client
+}
+
+func newRestHelper(httpClient *http.Client) *restHelper {
+	return &restHelper{httpClient: httpClient}
+}
+
+func (r *restHelper) deleteByIDWithSling(sling *sling.Sling, resourceID string, resourceDescription string) error {
+	request, _ := sling.New().Delete(resourceID).Request()
+	response, err := r.httpClient.Do(request)
+	if err != nil {
+		return err
+	}
+
+	defer response.Body.Close()
+	if response.StatusCode == 204 {
+		return nil
+	}
+
+	if response.StatusCode != 200 && response.StatusCode != 204 {
+		respData, err := ioutil.ReadAll(response.Body)
+		if err != nil {
+			return err
+		}
+		return fmt.Errorf("Error '%d' when deleting %s: %s", response.StatusCode, resourceDescription, string(respData))
+	}
+
+	return nil
+}
