+++ pkg/teamcity/vcs_trigger_options_test.go
@@ -112,29 +112,3 @@ func Test_VcsTriggerOptions_EnablePerCheckinTriggeringDisablesQueueOptimization(
 
 	props.assertPropertyValue(actual.vcsTriggerProperties(), "perCheckinTriggering", "true")
 }
-
-type PropertyAssertions struct {
-	a *assert.Assertions
-	t *testing.T
-}
-
-func newPropertyAssertions(t *testing.T) *PropertyAssertions {
-	return &PropertyAssertions{a: assert.New(t), t: t}
-}
-
-func (p *PropertyAssertions) assertPropertyValue(props *Properties, name string, value string) {
-	require.NotNil(p.t, props)
-
-	propMap := props.Map()
-
-	p.a.Contains(propMap, name)
-	p.a.Equal(value, propMap[name])
-}
-
-func (p *PropertyAssertions) assertPropertyDoesNotExist(props *Properties, name string) {
-	require.NotNil(p.t, props)
-
-	propMap := props.Map()
-
-	p.a.NotContains(propMap, name)
-}
